dmm-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
motor==3.3.2
pydantic==1.10.14
python-jose==3.3.0
loguru==0.7.2